Allow choices for Signal.signal to be configured
This adds a DB_MAILER_SIGNAL_CHOICES setting.  It also changes signal
values from expecting to be a name for a member in the module
`django.db.models.signals` and now expects a full path to a signal.
This allows user to specify other signals that they may choose to
define.  The migrations will rename all records in the database that
uses the previous naming convention.